+++ testing/testing/unusual_fields/models.py
@@ -20,9 +20,11 @@ import django
 
 from django.db import models
 from django.contrib.auth.models import User
+from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import ugettext_lazy as _
 
 
+@python_2_unicode_compatible
 class UnusualModel(models.Model):
 
     comma_field = models.CommaSeparatedIntegerField(max_length=250)
@@ -56,5 +58,5 @@ class UnusualModel(models.Model):
     def get_absolute_url(self):
         return ('unusual_edit', (self.pk,))
 
-    def __unicode__(self):
+    def __str__(self):
         return self.slug_field