+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Releases
 ========
 
+1.2.1 (2013-09-??)
+------------------
+
+* New options fieldTypes, and focusWhenEditing
+* Fix some details of getting the value, now there is a getValue function
+* Python 2/3 unicode compatible in testing project
+
 1.2.0 (2013-09-10)
 ------------------